add functionality for only logged in users to add, edit and delete notes

+++ notes.py
@@ -145,12 +145,12 @@ def logout():
     return redirect(url_for('index'))
 
 @app.route('/notes', methods=['GET', 'POST'])
-@login_required
 def notes():
     notes = Note.query.all()
     return render_template('notes.html', notes=notes)
 
 @app.route('/add_note', methods=['GET', 'POST'])
+@login_required
 def add_note():
     form = NoteForm()
     if form.validate_on_submit():
@@ -163,6 +163,7 @@ def add_note():
     return render_template('add_note.html', form=form)
 
 @app.route('/notes/edit/<int:id>', methods=['GET', 'POST'])
+@login_required
 def edit_note(id):
     note = Note.query.get_or_404(id)
     form = NoteForm()
@@ -173,20 +174,32 @@ def edit_note(id):
         db.session.commit()
         flash('The note has been updated')
         return redirect(url_for('.notes', id=note.id))
-    form.title.data = note.title
-    form.content.data = note.content
-    return render_template('edit_note.html', form=form)
-
-@app.route('/notes/delete/<int:id>', methods=['GET', 'POST'])
-def delete_note(id):
-    note_to_delete = Note.query.get_or_404(id)
-    try:
-        db.session.delete(note_to_delete)
-        db.session.commit()
-        flash('Note was deleted')
+    if current_user.id == note.user_id:
+        form.title.data = note.title
+        form.content.data = note.content
+        return render_template('edit_note.html', form=form)
+    else:
+        flash("You don't have a permission to edit this note")
         notes = Note.query.all()
         return render_template('notes.html', notes=notes)
-    except: 
-        flash('There was a problem deleting note. Try again')
+
+@app.route('/notes/delete/<int:id>', methods=['GET', 'POST'])
+@login_required
+def delete_note(id):
+    note_to_delete = Note.query.get_or_404(id)
+    id = current_user.id
+    if id == note_to_delete.user_id:
+        try:
+            db.session.delete(note_to_delete)
+            db.session.commit()
+            flash('Note was deleted')
+            notes = Note.query.all()
+            return render_template('notes.html', notes=notes)
+        except: 
+            flash('There was a problem deleting note. Try again')
+            notes = Note.query.all()
+            return render_template('notes.html', notes=notes)
+    else:
+        flash("You don't have a permission to delete this note")
         notes = Note.query.all()
         return render_template('notes.html', notes=notes)
